+++ src/high/job/mod.rs
@@ -1,15 +1,16 @@
-use crate::low::{vkcmd, vkdescriptor, vkmem, vkpipeline, vkshader, vkstate};
+use crate::low::{vkcmd, vkdescriptor, vkfence, vkmem, vkpipeline, vkshader, vkstate};
 use crate::utils::get_fract_s;
 pub use ash::version::{DeviceV1_0, EntryV1_0, InstanceV1_0};
 use ash::vk;
+use std::cell::RefCell;
 use std::ffi::CString;
 use std::fmt;
 use std::path::PathBuf;
+use std::rc::Rc;
 use std::time::Duration;
 use std::time::Instant;
 
-pub struct Timings {
-    pub init: Duration,
+pub struct JobTimings {
     pub upload: Duration,
     pub shader: Duration,
     pub cmd: Duration,
@@ -17,9 +18,8 @@ pub struct Timings {
     pub download: Duration,
 }
 
-pub struct TimingsBuilder {
-    init_timer: Option<Instant>,
-    init: Option<Duration>,
+#[derive(Debug, Clone, Copy)]
+pub struct JobTimingsBuilder {
     upload_timer: Option<Instant>,
     upload: Option<Duration>,
     shader_timer: Option<Instant>,
@@ -32,11 +32,9 @@ pub struct TimingsBuilder {
     download: Option<Duration>,
 }
 
-impl TimingsBuilder {
-    pub fn new() -> TimingsBuilder {
-        TimingsBuilder {
-            init_timer: None,
-            init: None,
+impl JobTimingsBuilder {
+    pub fn new() -> JobTimingsBuilder {
+        JobTimingsBuilder {
             upload_timer: None,
             upload: None,
             shader_timer: None,
@@ -50,81 +48,69 @@ impl TimingsBuilder {
         }
     }
 
-    pub fn start_init(mut self) -> TimingsBuilder {
-        self.init_timer = Some(Instant::now());
-        self
-    }
-
-    pub fn stop_init(mut self) -> TimingsBuilder {
-        self.init = self.init_timer.map(|instant| instant.elapsed());
-        self
-    }
-
-    pub fn start_upload(mut self) -> TimingsBuilder {
+    pub fn start_upload(mut self) -> JobTimingsBuilder {
         self.upload_timer = Some(Instant::now());
         self
     }
 
-    pub fn stop_upload(mut self) -> TimingsBuilder {
+    pub fn stop_upload(mut self) -> JobTimingsBuilder {
         self.upload = self.upload_timer.map(|instant| instant.elapsed());
         self
     }
 
-    pub fn start_shader(mut self) -> TimingsBuilder {
+    pub fn start_shader(mut self) -> JobTimingsBuilder {
         self.shader_timer = Some(Instant::now());
         self
     }
 
-    pub fn stop_shader(mut self) -> TimingsBuilder {
+    pub fn stop_shader(mut self) -> JobTimingsBuilder {
         self.shader = self.shader_timer.map(|instant| instant.elapsed());
         self
     }
 
-    pub fn start_cmd(mut self) -> TimingsBuilder {
+    pub fn start_cmd(mut self) -> JobTimingsBuilder {
         self.cmd_timer = Some(Instant::now());
         self
     }
 
-    pub fn stop_cmd(mut self) -> TimingsBuilder {
+    pub fn stop_cmd(mut self) -> JobTimingsBuilder {
         self.cmd = self.cmd_timer.map(|instant| instant.elapsed());
         self
     }
 
-    pub fn start_execution(mut self) -> TimingsBuilder {
+    pub fn start_execution(mut self) -> JobTimingsBuilder {
         self.execution_timer = Some(Instant::now());
         self
     }
 
-    pub fn stop_execution(mut self) -> TimingsBuilder {
+    pub fn stop_execution(mut self) -> JobTimingsBuilder {
         self.execution = self.execution_timer.map(|instant| instant.elapsed());
         self
     }
 
-    pub fn start_download(mut self) -> TimingsBuilder {
+    pub fn start_download(mut self) -> JobTimingsBuilder {
         self.download_timer = Some(Instant::now());
         self
     }
 
-    pub fn stop_download(mut self) -> TimingsBuilder {
+    pub fn stop_download(mut self) -> JobTimingsBuilder {
         self.download = self.download_timer.map(|instant| instant.elapsed());
         self
     }
 
-    pub fn build(self) -> Timings {
-        Timings {
-            init: self.init.unwrap(),
-            upload: self.upload.unwrap(),
-            shader: self.shader.unwrap(),
-            cmd: self.cmd.unwrap(),
-            execution: self.execution.unwrap(),
-            download: self.download.unwrap(),
+    pub fn build(self) -> JobTimings {
+        JobTimings {
+            upload: self.upload.unwrap_or_default(),
+            shader: self.shader.unwrap_or_default(),
+            cmd: self.cmd.unwrap_or_default(),
+            execution: self.execution.unwrap_or_default(),
+            download: self.download.unwrap_or_default(),
         }
     }
 }
 
-impl fmt::Display for Timings {
+impl fmt::Display for JobTimings {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "init: {}ms\n", get_fract_s(self.init))?;
         write!(f, "upload: {}ms\n", get_fract_s(self.upload))?;
         write!(f, "shader: {}ms\n", get_fract_s(self.shader))?;
         write!(f, "command: {}ms\n", get_fract_s(self.cmd))?;
@@ -133,9 +119,7 @@ impl fmt::Display for Timings {
         write!(
             f,
             "total: {}ms\n",
-            get_fract_s(
-                self.init + self.upload + self.shader + self.cmd + self.execution + self.download
-            )
+            get_fract_s(self.upload + self.shader + self.cmd + self.execution + self.download)
         )
     }
 }
@@ -152,11 +136,28 @@ impl BindPoint {
     }
 }
 
+#[derive(Eq, PartialEq, Debug)]
+pub enum JobStatus {
+    INIT,
+    EXECUTING,
+    SUCESS,
+    FAILURE,
+}
+
 pub struct Job<'a, T> {
     inputs: Vec<(BindPoint, &'a Vec<T>)>,
     buffers: Vec<(BindPoint, usize)>,
     shaders: Vec<&'a PathBuf>,
     dispatch: Vec<(u32, u32, u32)>,
+    state: JobState,
+}
+
+pub struct JobState {
+    timing: JobTimingsBuilder,
+    fence: Option<vkfence::VkFence>,
+    memory: Option<vkmem::VkMem>,
+    buffers: Vec<vkmem::VkBuffer>,
+    vulkan: Rc<vkstate::VulkanState>,
 }
 
 pub struct JobBuilder<'a, T> {
@@ -196,12 +197,20 @@ impl<'a, T> JobBuilder<'a, T> {
         self
     }
 
-    pub fn build(self) -> Job<'a, T> {
+    pub fn build(self, vulkan: Rc<vkstate::VulkanState>) -> Job<'a, T> {
+        let state = JobState {
+            fence: None,
+            buffers: Vec::new(),
+            memory: None,
+            timing: JobTimingsBuilder::new(),
+            vulkan: vulkan,
+        };
         Job {
             inputs: self.inputs,
             buffers: self.buffers,
             shaders: self.shaders,
             dispatch: self.dispatch,
+            state: state,
         }
     }
 }
@@ -209,21 +218,14 @@ impl<'a, T> JobBuilder<'a, T> {
 // TODO: Correctly manage set binding.
 // For the moment, all binding will use the set 0 neverminding the actual value in BindPoint
 impl<'a, T> Job<'a, T> {
-    pub fn execute(&self) -> (Vec<Vec<T>>, Timings) {
-        let mut timing_builder = TimingsBuilder::new();
+    pub fn execute(&mut self) {
         let inputs = &self.inputs;
         let ro_buffers = &self.buffers;
         let shaders = &self.shaders;
         let dispatch = &self.dispatch;
 
-        // Vulkan init.
-        timing_builder = timing_builder.start_init();
-        let vulkan = vkstate::init_vulkan();
-        vkstate::print_work_limits(&vulkan);
-        timing_builder = timing_builder.stop_init();
-
         // Memory init.
-        timing_builder = timing_builder.start_upload();
+        self.state.timing = self.state.timing.start_upload();
         let mut buffer_sizes: Vec<u64> = inputs
             .iter()
             .map(|v| (v.1.len() * std::mem::size_of::<T>()) as u64)
@@ -232,30 +234,34 @@ impl<'a, T> Job<'a, T> {
             buffer_sizes.push((s.1 * std::mem::size_of::<T>()) as u64);
         }
 
-        let mut buffers: Vec<vkmem::VkBuffer> = buffer_sizes
+        self.state.buffers = buffer_sizes
             .iter()
-            .map(|size| vkmem::VkBuffer::new(&vulkan, *size))
+            .map(|size| vkmem::VkBuffer::new(self.state.vulkan.clone(), *size))
             .collect();
-        let (mem_size, offsets) = vkmem::compute_non_overlapping_buffer_alignment(&buffers);
-        let vk_mem = vkmem::VkMem::find_mem(&vulkan, mem_size);
-        if vk_mem.is_none() {
-            panic!("[ERR] Could not find a memory type fitting our need.");
-        }
+        let (mem_size, offsets) =
+            vkmem::compute_non_overlapping_buffer_alignment(&self.state.buffers);
+        self.state.memory = Some(
+            vkmem::VkMem::find_mem(self.state.vulkan.clone(), mem_size)
+                .expect("[ERR] Could not find a memory type fitting our need."),
+        );
 
-        let vk_mem = vk_mem.unwrap();
-        for i in 0..buffers.len() {
-            let mbuf = buffers.get_mut(i).unwrap();
-            mbuf.bind(vk_mem.mem, offsets[i]);
+        for i in 0..self.state.buffers.len() {
+            let mbuf = self.state.buffers.get_mut(i).unwrap();
+            mbuf.bind(self.state.memory.as_ref().unwrap().mem, offsets[i]);
             if i < inputs.len() {
-                vk_mem.map_buffer(inputs[i].1, mbuf);
+                self.state
+                    .memory
+                    .as_ref()
+                    .unwrap()
+                    .map_buffer(inputs[i].1, mbuf);
             }
         }
 
-        timing_builder = timing_builder.stop_upload();
+        self.state.timing = self.state.timing.stop_upload();
 
         // Shaders
-        timing_builder = timing_builder.start_shader();
-        let mut shad_vec: Vec<vkshader::VkShader> = Vec::with_capacity(shaders.len());
+        self.state.timing = self.state.timing.start_shader();
+        let mut shad_vec: Vec<Rc<RefCell<vkshader::VkShader>>> = Vec::with_capacity(shaders.len());
         let mut shad_pip_vec: Vec<vkpipeline::VkComputePipeline> =
             Vec::with_capacity(shaders.len());
         let mut shad_pipeline_layout: Vec<vk::PipelineLayout> = Vec::with_capacity(shaders.len());
@@ -263,30 +269,41 @@ impl<'a, T> Job<'a, T> {
         let mut shad_desc_set: Vec<vkdescriptor::VkWriteDescriptor> =
             Vec::with_capacity(shaders.len());
         for path in shaders {
-            shad_vec.push(vkshader::VkShader::new(
-                &vulkan,
+            shad_vec.push(Rc::new(RefCell::new(vkshader::VkShader::new(
+                self.state.vulkan.clone(),
                 path,
                 CString::new("main").unwrap(),
-            ));
+            ))));
         }
         for shader in shad_vec.iter_mut() {
-            for i in 0..buffers.len() {
-                shader.add_layout_binding(
+            for i in 0..self.state.buffers.len() {
+                shader.borrow_mut().add_layout_binding(
                     i as u32,
                     1,
                     vk::DescriptorType::STORAGE_BUFFER,
                     vk::ShaderStageFlags::COMPUTE,
                 );
             }
-            shader.create_pipeline_layout();
-            shad_pipeline_layout.push(shader.pipeline.unwrap());
-            shad_pip_vec.push(vkpipeline::VkComputePipeline::new(&vulkan, shader));
-            shad_desc_vec.push(vkdescriptor::VkDescriptor::new(&vulkan, shader));
-            shad_desc_set.push(vkdescriptor::VkWriteDescriptor::new(&vulkan));
+            shader.borrow_mut().create_pipeline_layout();
+            shad_pipeline_layout.push(shader.borrow().pipeline.unwrap());
+            shad_pip_vec.push(vkpipeline::VkComputePipeline::new(
+                self.state.vulkan.clone(),
+                &shader.borrow(),
+            ));
+            shad_desc_vec.push(vkdescriptor::VkDescriptor::new(
+                self.state.vulkan.clone(),
+                shader.clone(),
+            ));
+            shad_desc_set.push(vkdescriptor::VkWriteDescriptor::new(
+                self.state.vulkan.clone(),
+            ));
         }
 
         for descriptor in shad_desc_vec.iter_mut() {
-            descriptor.add_pool_size(buffers.len() as u32, vk::DescriptorType::STORAGE_BUFFER);
+            descriptor.add_pool_size(
+                self.state.buffers.len() as u32,
+                vk::DescriptorType::STORAGE_BUFFER,
+            );
             descriptor.create_pool(1);
             descriptor.create_set();
         }
@@ -295,8 +312,12 @@ impl<'a, T> Job<'a, T> {
         for write_descriptor_set in shad_desc_set.iter_mut() {
             let desc_set: vk::DescriptorSet = *shad_desc_vec[n].get_first_set().unwrap();
             let mut buffers_nfos: Vec<Vec<vk::DescriptorBufferInfo>> = Vec::new();
-            for i in 0..buffers.len() {
-                write_descriptor_set.add_buffer(buffers[i].buffer, 0, buffers[i].size);
+            for i in 0..self.state.buffers.len() {
+                write_descriptor_set.add_buffer(
+                    self.state.buffers[i].buffer,
+                    0,
+                    self.state.buffers[i].size,
+                );
                 buffers_nfos.push(vec![write_descriptor_set.buffer_descriptors[i]]);
                 write_descriptor_set.add_write_descriptors(
                     desc_set,
@@ -310,12 +331,12 @@ impl<'a, T> Job<'a, T> {
             n += 1;
         }
 
-        timing_builder = timing_builder.stop_shader();
+        self.state.timing = self.state.timing.stop_shader();
 
         // Command buffers
-        timing_builder = timing_builder.start_cmd();
+        self.state.timing = self.state.timing.start_cmd();
         let mut cmd_buffers: Vec<usize> = Vec::with_capacity(shaders.len());
-        let mut cmd_pool = vkcmd::VkCmdPool::new(&vulkan);
+        let mut cmd_pool = vkcmd::VkCmdPool::new(self.state.vulkan.clone());
 
         for _ in 0..shaders.len() {
             cmd_buffers.push(cmd_pool.create_cmd_buffer(vk::CommandBufferLevel::PRIMARY));
@@ -336,7 +357,7 @@ impl<'a, T> Job<'a, T> {
 
             // Memory barrier
             let mut buffer_barrier: Vec<vk::BufferMemoryBarrier> = Vec::new();
-            for buffer in &buffers {
+            for buffer in &self.state.buffers {
                 buffer_barrier.push(
                     vk::BufferMemoryBarrier::builder()
                         .src_access_mask(vk::AccessFlags::SHADER_WRITE)
@@ -347,7 +368,7 @@ impl<'a, T> Job<'a, T> {
                 );
             }
             unsafe {
-                vulkan.device.cmd_pipeline_barrier(
+                self.state.vulkan.device.cmd_pipeline_barrier(
                     cmd_pool.cmd_buffers[i],
                     vk::PipelineStageFlags::COMPUTE_SHADER,
                     vk::PipelineStageFlags::COMPUTE_SHADER,
@@ -360,26 +381,59 @@ impl<'a, T> Job<'a, T> {
 
             cmd_pool.end_cmd(i);
         }
-        timing_builder = timing_builder.stop_cmd();
+        self.state.timing = self.state.timing.stop_cmd();
 
         // Execution
-        timing_builder = timing_builder.start_execution();
-        let queue = unsafe { vulkan.device.get_device_queue(vulkan.queue_family_index, 0) };
-        cmd_pool.submit(queue);
-
-        unsafe {
-            vulkan
+        self.state.fence = Some(vkfence::VkFence::new(self.state.vulkan.clone(), false));
+        self.state.timing = self.state.timing.start_execution();
+        let queue = unsafe {
+            self.state
+                .vulkan
                 .device
-                .queue_wait_idle(queue)
-                .expect("[ERR] Error while waiting for queue to be idle.")
+                .get_device_queue(self.state.vulkan.queue_family_index, 0)
         };
-        timing_builder = timing_builder.stop_execution();
+        cmd_pool.submit(queue, Some(self.state.fence.as_ref().unwrap().fence));
+    }
 
-        // Download results.
-        timing_builder = timing_builder.start_download();
-        let output: Vec<Vec<T>> = buffers.iter().map(|buf| vk_mem.get_buffer(buf)).collect();
-        timing_builder = timing_builder.stop_download();
+    pub fn status(&self) -> JobStatus {
+        if self.state.fence.is_none() {
+            return JobStatus::INIT;
+        }
 
-        (output, timing_builder.build())
+        let r = self.state.fence.as_ref().map(|fence| fence.status());
+        match r.unwrap() {
+            vkfence::FenceStates::SIGNALED => JobStatus::SUCESS,
+            vkfence::FenceStates::UNSIGNALED => JobStatus::EXECUTING,
+            _ => JobStatus::FAILURE,
+        }
+    }
+
+    pub fn get_output(&self) -> Option<Vec<Vec<T>>> {
+        if self.status() != JobStatus::SUCESS {
+            return None;
+        }
+
+        let output: Vec<Vec<T>> = self
+            .state
+            .buffers
+            .iter()
+            .map(|buf| self.state.memory.as_ref().unwrap().get_buffer(buf))
+            .collect();
+
+        Some(output)
+    }
+
+    pub fn get_timing(&self) -> JobTimings {
+        self.state.timing.build()
+    }
+
+    pub fn wait_until_idle(&self, timeout: u64) -> JobStatus {
+        let current_status = self.status();
+        if current_status != JobStatus::EXECUTING {
+            return current_status;
+        }
+
+        self.state.fence.as_ref().map(|fence| fence.wait(timeout));
+        self.status()
     }
 }