+++ src/low/mod.rs
@@ -1,5 +1,6 @@
 pub mod vkcmd;
 pub mod vkdescriptor;
+pub mod vkfence;
 pub mod vkmem;
 pub mod vkpipeline;
 pub mod vkshader;